finish auth with status

+++ controllers/feed.js
@@ -3,7 +3,7 @@ const fs = require('fs');
 const path = require('path');
 
 const Post = require('../models/post');
-
+const User = require('../models/user');
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
   const perPage = 2;
@@ -39,7 +39,7 @@ exports.createPost = (req, res, next) => {
     throw error;
   }
   const { title, content } = req.body;
-
+  let creator;
   if (!req.file) {
     const error = new Error('No image provided');
     error.statusCode = 422;
@@ -50,17 +50,28 @@ exports.createPost = (req, res, next) => {
   const post = new Post({
     title: title,
     content: content,
-    creator: { name: 'Leonardo' },
+    creator: req.userId,
     imageUrl: imageUrl,
   });
   //create post in db
   post
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: 'Post created succesfully',
-        post: result,
+        post: post,
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -121,6 +132,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!.');
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -148,17 +164,80 @@ exports.deletePost = (req, res, next) => {
   Post.findById(postId)
     .then((post) => {
       //Check logged in user
-      clearImage(post.imageUrl);
       if (!post) {
         const error = new Error('Could not find post.');
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authorized!.');
+        error.statusCode = 403;
+        throw error;
+      }
+      clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: 'Deleted post!' });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.getStatus = (req, res, next) => {
+  User.findById(req.userId)
+    .then((user) => {
+      if (!user) {
+        const error = new Error("Can't find the user");
+        error.statusCode = 403;
+        throw error;
+      }
+      res.status(200).json({
+        status: user.status,
+      });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.patchStatus = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const error = new Error('Validation Failed, entered data is incorrect.');
+    error.statusCode = 422;
+    throw error;
+  }
+  const { status } = req.body;
+
+  User.findById(req.userId)
+    .then((user) => {
+      if (!user) {
+        const error = new Error('User not found');
+        error.statusCode = 422;
+        throw error;
+      }
+      user.status = status;
+      return user.save();
+    })
+    .then((result) => {
+      res.status(200).json({
+        message: 'User Updated.',
+      });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -170,7 +249,6 @@ exports.deletePost = (req, res, next) => {
 
 const clearImage = (filePath) => {
   filePath = path.join(__dirname, '..', filePath);
-  console.log(filePath);
   fs.unlink(filePath, (err) => {
     if (err) {
       const error = new Error('Could not delete image.');
